Changing file structure and converting more of example sentences to object oriented
