+++ app/mult_words_blanks.py
@@ -10,6 +10,3 @@ class All_words:
             six_words = six_words_with_spaces.Six_words_with_blanks()
             self.all_words += six_words.generate_six_words_with_blanks(word) + "\n"
         return self.all_words
-
-all_words = All_words()
-print(all_words.generate_all_words(["apple", "carrot"]))
