+++ sentences.txt
@@ -188,3 +188,4 @@ A cyclops has one eye.
 Gold is worth a lot of money.
 My father works during the day.
 There is grass on the field. 
+"There is an apple in the fridge"
